pyproject.toml
[project]
name = "lifebot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
python-dotenv = ">=1.1.0,<2.0.0"
python-telegram-bot = {extras = ["job-queue"], version = ">=22.1,<23.0"}
redis = ">=6.2.0,<7.0.0"
sqlalchemy = ">=2.0.41,<3.0.0"
requests = ">=2.32.4,<3.0.0"
loguru = ">=0.7.3,<0.8.0"
greenlet = "^3.2.3"
pydantic = {extras = ["mypy"], version = "^2.11.7"}
aiomysql = "^0.2.0"
cryptography = "^45.0.4"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
autoflake = "^2.3.1"
types-requests = "^2.32.4.20250611"
black = "^25.1.0"
mypy = "^1.16.1"

[tool.black]
line-length=88
extend-exclude = "/(venv|.venv)/"
target-version = ["py313"]

[tool.isort]
line_length=88
profile="black"
skip_gitignore=true
skip = ["venv", ".venv"]
combine_as_imports = true
include_trailing_comma = true
multi_line_output = 3
ensure_newline_before_comments = true

[tool.mypy]
plugins = ['pydantic.mypy']
python_version = "3.13"
strict = true
mypy_path = ["src"]
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.autoflake]
remove_all_unused_imports = true
remove_unused_variables = true
recursive = true
in_place = true